解析弹窗

+++ app/src/main/java/com/app/videobox/ui/widgets/VideoCardV1.kt
@@ -116,7 +116,7 @@ fun VideoCardV1(
                     }
                     VideoTimeInfoLabel(
                         modifier = Modifier.align(Alignment.BottomEnd),
-                        duration = duration,
+                        duration = duration.toInt(),
                     )
 
                 }
